+++ src/main/kotlin/com/prestongarno/ktq/stubs/ScalarArrays.kt
@@ -17,12 +17,10 @@
 
 package com.prestongarno.ktq.stubs
 
-import com.prestongarno.ktq.ArgBuilder
 import com.prestongarno.ktq.QModel
 import com.prestongarno.ktq.SchemaStub
 import com.prestongarno.ktq.adapters.Adapter
-import com.prestongarno.ktq.adapters.bind
-import com.prestongarno.ktq.adapters.formatAs
+import com.prestongarno.ktq.internal.formatAs
 import com.prestongarno.ktq.internal.CollectionDelegate
 import com.prestongarno.ktq.properties.GraphQlProperty
 import kotlin.reflect.KProperty
