+++ src/main/kotlin/com/prestongarno/ktq/stubs/Scalars.kt
@@ -20,7 +20,7 @@ package com.prestongarno.ktq.stubs
 import com.prestongarno.ktq.QModel
 import com.prestongarno.ktq.SchemaStub
 import com.prestongarno.ktq.adapters.Adapter
-import com.prestongarno.ktq.adapters.formatAs
+import com.prestongarno.ktq.internal.formatAs
 import com.prestongarno.ktq.internal.ValueDelegate
 import com.prestongarno.ktq.properties.GraphQlProperty
 import kotlin.reflect.KProperty
